+++ src/main/java/com/battleclub/gameserver/GameUpdateLoop.java
@@ -11,6 +11,12 @@ import org.springframework.scheduling.annotation.Scheduled;
 import com.battleclub.gameserver.Game.Coordinates;
 import com.battleclub.gameserver.Game.GameState;
 import com.battleclub.gameserver.Game.Player;
+import com.battleclub.gameserver.player.PlayerInputManager;
+import com.battleclub.gameserver.player.PlayerInputMessage;
+import com.battleclub.gameserver.player.PlayerMovement;
+import com.battleclub.user.Users;
+import com.battleclub.utils.CoordinateUtils;
+import com.battleclub.utils.WeaponSelection;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
