+++ src/main/java/com/battleclub/gameserver/GameserverApplication.java
@@ -3,7 +3,7 @@ package com.battleclub.gameserver;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {".", "player","user","utils"})
 public class GameserverApplication {
 
 	public static void main(String[] args) {
