+++ src/main/java/com/battleclub/gameserver/player/PlayerJoinInterceptor.java
@@ -1,4 +1,4 @@
-package com.battleclub.gameserver;
+package com.battleclub.gameserver.player;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.Message;
@@ -6,6 +6,8 @@ import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.simp.stomp.StompCommand;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.messaging.support.ChannelInterceptor;
+
+import com.battleclub.gameserver.Game;
 
 public class PlayerJoinInterceptor implements ChannelInterceptor {
     @Autowired
