+++ src/main/java/com/battleclub/user/UserIdController.java
@@ -1,9 +1,11 @@
-package com.battleclub.gameserver;
+package com.battleclub.user;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.stereotype.Controller;
+
+import com.battleclub.gameserver.Game;
 
 import lombok.extern.slf4j.Slf4j;
