+++ src/main/java/com/battleclub/utils/RandomNameGenerator.java
@@ -1,4 +1,4 @@
-package com.battleclub.gameserver;
+package com.battleclub.utils;
 
 import java.util.Random;
